requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
tensorflow
opencv-python
ultralytics
numpy
pillow
python-multipart